pyproject.toml
[tool.poetry]
name = "test4"
version = "0.1.0"
description = ""
authors = ["PeretEtienne <etienne.peret@integragen.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
ruff = "^0.1.4"
mypy = "^1.6.1"
pyvcf3 = "^1.0.3"
elasticsearch = {extras = ["async"], version = "^8.11.0"}
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

# for strict mypy: (this is the tricky one :-))
disallow_untyped_defs = false

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
